Add files via upload
The study examines the macroeconomic indicators of economic growth in Ghana
measured by Gross Domestic Product(GDP). We used economic data from BoG
and GSS for a period of 36 years that is 1985 to 2021. Multiple Linear Regresson was used to establish a relationship between economic growth and its
macroeconomic indicators. T-test was used to test zero means of residuals and
Shapiro-Wilk test to check for normality of the error term of residuals. Studentizeed Breush-Pegan test was used to check for constant variance. The results
show that exchange rate, export and balance of trade are the main indicators
that have impact on Ghana’s economic growth under a stable inflationary rate.
The study recommended that there is a need to increase export, strengthen the
currency and reduce the trade deficit.